Move notification snackbar to separate feature and only clear on timeout
